pumpix: fix breaking of dimension check
previously, when pumpix configuration was in app.py
we relied on a direct tuple IMG_MAX_DIMENSIONS to compare
against <image>.size, which gives a (width, height) tuple

after moving to reading pumpixes configuration from a file,
we encounter an issue in that pyaml cannot support tuples in
the file without discarding safe loading which is undesirable.

the app was previously attempting to check (w, h) against a int,
which clearly breaks, this patch fixes this by introducing
MAX_IMAGE_WIDTH and MAX_IMAGE_HEIGHT, while deprecating
IMG_MAX_DIMENSIONS and masqeurading the two values
a tuple for the check.